refactor: Rearrange the structure of project 'HellSolver'
