+++ Includes/HellSolver/HellSolver.hpp
@@ -11,10 +11,10 @@
 #include <HellSolver/Agents/RandomAgent.hpp>
 #include <HellSolver/Enums/GameEnums.hpp>
 #include <HellSolver/Games/Game.hpp>
-#include <HellSolver/Games/Map.hpp>
-#include <HellSolver/Games/Object.hpp>
-#include <HellSolver/Games/Position.hpp>
 #include <HellSolver/HellSolver.hpp>
-#include <HellSolver/Player/Player.hpp>
+#include <HellSolver/Models/Map.hpp>
+#include <HellSolver/Models/Object.hpp>
+#include <HellSolver/Models/Player.hpp>
+#include <HellSolver/Models/Position.hpp>
 
 #endif  // HELL_SOLVER_HPP
