+++ Sources/HellSolver/Models/Object.cpp
@@ -4,7 +4,7 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <HellSolver/Games/Object.hpp>
+#include <HellSolver/Models/Object.hpp>
 
 #include <utility>
